Fix several bugs on mantis functions

+++ R/mantis.R
@@ -11,7 +11,8 @@
 #' @examples
 #' scenario<-mantis.ScenarioOptions()
 mantis.ScenarioOptions <- function(){
-  opt <-list(infile = 'incomingMSG.dat',
+  opt <-list( client = NA,
+              infile = 'incomingMSG.dat',
               outfile = 'testClientResults.dat',
               descr = 'This is a description of the simulation input',
               endSimYear = 2100,
@@ -43,21 +44,27 @@ mantis.ScenarioOptions <- function(){
 #' mantis.Run Runs the scenario and returns the results
 #'
 #' @param scenario is a list with the scenario options
-#' @param client_prog is the path to the Mantis client executable
 #'
 #' @return a list with the well breakthrough curves
 #' @export
 #'
 #' @examples
-#' result <- mantis.Run(scenario, 'MantisClient.exe')
-mantis.Run <- function(scenario, client_prog){
+#' scenario <- mantis.ScenarioOptions()
+#' scenario$client <- 'MantisClient.exe'
+#' result <- mantis.Run(scenario)
+mantis.Run <- function(scenario){
+  if (is.na(scenario$client)){
+    outmsg <- 'Client Path not specified'
+    return(outmsg)
+  }
+
   # Remove the output file
   if (file.exists(scenario$outfile)){
     file.remove(scenario$outfile)
   }
   mantis.WriteInput(scenario)
 
-  system(paste(client_prog, scenario$infile, scenario$outfile))
+  system(paste(scenario$client, scenario$infile, scenario$outfile))
 
   res <- mantis.ReadOutput(scenario$outfile)
   return(res)
@@ -66,15 +73,49 @@ mantis.Run <- function(scenario, client_prog){
 
 #' mantis.Quit Stops the server
 #'
-#' @param client_prog is the path to the Mantis client executable
+#' @param scenario the only useful option for this function
+#' is the client which is is the path to the client executable
 #'
 #' @return Returns nothing. It just quits the server
 #' @export
 #'
 #' @examples
-#' mantis.Quit <- function( 'MantisClient.exe')
-mantis.Quit <- function(client_prog){
-  system(paste(client_prog, 'quit'))
+#' scenario <- mantis.ScenarioOptions()
+#' scenario$client <- 'MantisClient.exe'
+#' mantis.Quit <- function( scenario)
+mantis.Quit <- function(scenario){
+  if (!is.na(scenario$client)){
+    system(paste(scenario$client, 'quit'))
+    return(1)
+  }
+  else{
+    outmsg <- 'Client Path not specified'
+    return(outmsg)
+  }
+}
+
+#' mantis.Ping Test the communication with server.
+#' It sends a ping and expects a pong
+#'
+#' @param scenario the only useful option for this function
+#' is the client which is is the path to the client executable
+#'
+#' @return Returns a pong if the communication is succesful
+#' @export
+#'
+#' @examples
+#' scenario <- mantis.ScenarioOptions()
+#' scenario$client <- 'MantisClient.exe'
+#' mantis.Ping <- function( scenario)
+mantis.Ping <- function(scenario){
+  if (!is.na(scenario$client)){
+    system(paste(scenario$client, 'ping', scenario$outfile))
+    outmsg <- mantis.ReadOutput(scenario$outfile)
+  }
+  else{
+    outmsg <- 'Client Path not specified'
+  }
+  return(outmsg)
 }
 
 
@@ -100,7 +141,7 @@ mantis.WriteInput <-function(scenario){
     appnd <- TRUE
   }
 
-  for (i in 4:length(scenNames)) {
+  for (i in 5:length(scenNames)) {
     if (!is.na(scenario[[i]][1])){
       if (scenNames[i] == 'Regions'){
         tmp <- paste('Nregions', length(scenario$Regions))
@@ -143,8 +184,14 @@ mantis.ReadOutput <- function(filename){
   con <- file(filename, open = "r")
   line <- readLines(con,1)
   close(con)
-  t <- as.numeric(strsplit(substr(line,1,2000), split = " "))
+  t <- strsplit(substr(line,1,100), split = " ")
   NNs <- as.numeric(t[[1]])
-  res <- utils::read.table(file='filename',skip = 1)
-  return(res)
+  if (NNs[1] == 0){
+    alllines <- readLines(filename)
+    return(alllines)
+  }
+  else{
+    res <- utils::read.table(file = filename, skip = 1)
+    return(res)
+  }
 }
